Create Button as reusable UI
