+++ src/Components/ExperimentOne/ExperimentOne.js
@@ -1,20 +1,16 @@
 import ExperimentOneDisplayInformation from "./ExperimentOneDisplayInformation.js";
 import ExperimentOneHeader from "./ExperimentOneHeader.js";
+import Button from "../ReUsableUIs/Button.js";
 
 const ExperimentOne = ({ isBackButtonClicked }) => {
-  const backButtonStyle = {
-    fontSize: "20px",
-    backgroundColor: "black",
-    color: "white",
-    margin: "20px 500px",
-  };
   return (
     <>
       <ExperimentOneHeader />
       <ExperimentOneDisplayInformation />
-      <button style={backButtonStyle} onClick={isBackButtonClicked}>
-        Back to Main Arena
-      </button>
+      <Button
+        textToDisplay="Head to Main Arena"
+        onClickListener={isBackButtonClicked}
+      />
     </>
   );
 };
